# Allow GC to collect unneeded slice elements