--- a/sdk/metric/pipeline.go
+++ b/sdk/metric/pipeline.go
@@ -115,6 +115,7 @@ func (p *pipeline) produce(ctx context.Context, rm *metricdata.ResourceMetrics) 
 		}
 		if err := ctx.Err(); err != nil {
 			rm.Resource = nil
+			clear(rm.ScopeMetrics) // Erase elements to let GC collect objects
 			rm.ScopeMetrics = rm.ScopeMetrics[:0]
 			return err
 		}
@@ -128,6 +129,7 @@ func (p *pipeline) produce(ctx context.Context, rm *metricdata.ResourceMetrics) 
 		if err := ctx.Err(); err != nil {
 			// This means the context expired before we finished running callbacks.
 			rm.Resource = nil
+			clear(rm.ScopeMetrics) // Erase elements to let GC collect objects
 			rm.ScopeMetrics = rm.ScopeMetrics[:0]
 			return err
 		}