```go

type interInst struct {
	x int
}

type inter interface {
}

var sink []inter

func BenchmarkX(b *testing.B) {
	sink = make([]inter, b.N)
	for i := 0; i < b.N; i++ {
		sink[i] = &interInst{}
	}
	clear(sink)
	sink = sink[:0]
	runtime.GC()
	var ms runtime.MemStats
	runtime.ReadMemStats(&ms)
	b.Log(b.N, ms.Frees) // Frees is the cumulative count of heap objects freed.
}
```

```
clear:
    ioz_test.go:35: 1 589
    ioz_test.go:35: 100 711
    ioz_test.go:35: 10000 10729
    ioz_test.go:35: 1000000 1010750  <-- 1m+ freed
    ioz_test.go:35: 16076874 17087643
    ioz_test.go:35: 19514749 36602412
```
```
no clear:
    ioz_test.go:35: 1 585
    ioz_test.go:35: 100 606
    ioz_test.go:35: 10000 725
    ioz_test.go:35: 1000000 10745  <-- some "overhead" objects freed, not the slice.
    ioz_test.go:35: 16391445 1010765
    ioz_test.go:35: 21765238 17402230
```

This is documented at https://go.dev/wiki/SliceTricks:

> NOTE If the type of the element is a pointer or a struct with pointer fields, which need to be garbage collected, the above implementations of Cut and Delete have a potential memory leak problem: some elements with values are still referenced by slice a’s underlying array, just not “visible” in the slice. Because the “deleted” value is referenced in the underlying array, the deleted value is still “reachable” during GC, even though the value cannot be referenced by your code. If the underlying array is long-lived, this represents a leak.

... followed by examples of how zeroing out the slice elements solves the problem. This PR does the same.

---

`sdk/log`:

```
goos: darwin
goarch: arm64
pkg: go.opentelemetry.io/otel/sdk/log
                                  │   ./old.txt   │               ./new.txt               │
                                  │    sec/op     │    sec/op      vs base                │
BatchProcessorOnEmit-10              153.0n ± ∞ ¹    151.3n ± ∞ ¹       ~ (p=1.000 n=1) ²
Processor/Simple-10                  448.5n ± ∞ ¹    453.6n ± ∞ ¹       ~ (p=1.000 n=1) ²
Processor/Batch-10                   887.8n ± ∞ ¹    981.1n ± ∞ ¹       ~ (p=1.000 n=1) ²
Processor/SetTimestampSimple-10      444.8n ± ∞ ¹    444.6n ± ∞ ¹       ~ (p=1.000 n=1) ²
Processor/SetTimestampBatch-10       981.6n ± ∞ ¹    954.1n ± ∞ ¹       ~ (p=1.000 n=1) ²
Processor/AddAttributesSimple-10     498.0n ± ∞ ¹    472.1n ± ∞ ¹       ~ (p=1.000 n=1) ²
Processor/AddAttributesBatch-10      983.7n ± ∞ ¹   1048.0n ± ∞ ¹       ~ (p=1.000 n=1) ²
Processor/SetAttributesSimple-10     463.7n ± ∞ ¹    463.4n ± ∞ ¹       ~ (p=1.000 n=1) ²
Processor/SetAttributesBatch-10      985.2n ± ∞ ¹   1028.0n ± ∞ ¹       ~ (p=1.000 n=1) ²
LoggerNewRecord/5_attributes-10      197.9n ± ∞ ¹    203.8n ± ∞ ¹       ~ (p=1.000 n=1) ²
LoggerNewRecord/10_attributes-10     648.0n ± ∞ ¹    658.4n ± ∞ ¹       ~ (p=1.000 n=1) ²
LoggerEnabled-10                     5.307n ± ∞ ¹    5.360n ± ∞ ¹       ~ (p=1.000 n=1) ²
LoggerProviderLogger-10              547.2n ± ∞ ¹    554.1n ± ∞ ¹       ~ (p=1.000 n=1) ²
WalkAttributes/1_attributes-10       3.427n ± ∞ ¹    3.434n ± ∞ ¹       ~ (p=1.000 n=1) ²
WalkAttributes/10_attributes-10      3.428n ± ∞ ¹    3.442n ± ∞ ¹       ~ (p=1.000 n=1) ²
WalkAttributes/100_attributes-10     3.427n ± ∞ ¹    3.435n ± ∞ ¹       ~ (p=1.000 n=1) ²
WalkAttributes/1000_attributes-10    3.425n ± ∞ ¹    3.432n ± ∞ ¹       ~ (p=1.000 n=1) ²
SetAddAttributes/SetAttributes-10    79.45n ± ∞ ¹    81.11n ± ∞ ¹       ~ (p=1.000 n=1) ²
SetAddAttributes/AddAttributes-10    59.47n ± ∞ ¹    61.89n ± ∞ ¹       ~ (p=1.000 n=1) ²
SimpleProcessorOnEmit-10            0.5219n ± ∞ ¹   0.2676n ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                              87.87n          86.11n        -2.01%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                        │   ./old.txt   │            ./new.txt             │
                        │      B/s      │      B/s       vs base           │
BatchProcessorOnEmit-10   199.5Mi ± ∞ ¹   201.8Mi ± ∞ ¹  ~ (p=1.000 n=1) ²
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                                  │   ./old.txt   │               ./new.txt               │
                                  │     B/op      │     B/op       vs base                │
BatchProcessorOnEmit-10               0.000 ± ∞ ¹     0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
Processor/Simple-10                   416.0 ± ∞ ¹     416.0 ± ∞ ¹       ~ (p=1.000 n=1) ²
Processor/Batch-10                  1.177Ki ± ∞ ¹   1.213Ki ± ∞ ¹       ~ (p=1.000 n=1) ³
Processor/SetTimestampSimple-10       416.0 ± ∞ ¹     416.0 ± ∞ ¹       ~ (p=1.000 n=1) ²
Processor/SetTimestampBatch-10      1.201Ki ± ∞ ¹   1.205Ki ± ∞ ¹       ~ (p=1.000 n=1) ³
Processor/AddAttributesSimple-10      416.0 ± ∞ ¹     416.0 ± ∞ ¹       ~ (p=1.000 n=1) ²
Processor/AddAttributesBatch-10     1.195Ki ± ∞ ¹   1.211Ki ± ∞ ¹       ~ (p=1.000 n=1) ³
Processor/SetAttributesSimple-10      464.0 ± ∞ ¹     464.0 ± ∞ ¹       ~ (p=1.000 n=1) ²
Processor/SetAttributesBatch-10     1.251Ki ± ∞ ¹   1.263Ki ± ∞ ¹       ~ (p=1.000 n=1) ³
LoggerNewRecord/5_attributes-10       0.000 ± ∞ ¹     0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
LoggerNewRecord/10_attributes-10      609.0 ± ∞ ¹     609.0 ± ∞ ¹       ~ (p=1.000 n=1) ²
LoggerEnabled-10                      0.000 ± ∞ ¹     0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
LoggerProviderLogger-10               352.0 ± ∞ ¹     351.0 ± ∞ ¹       ~ (p=1.000 n=1) ³
WalkAttributes/1_attributes-10        0.000 ± ∞ ¹     0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
WalkAttributes/10_attributes-10       0.000 ± ∞ ¹     0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
WalkAttributes/100_attributes-10      0.000 ± ∞ ¹     0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
WalkAttributes/1000_attributes-10     0.000 ± ∞ ¹     0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
SetAddAttributes/SetAttributes-10     48.00 ± ∞ ¹     48.00 ± ∞ ¹       ~ (p=1.000 n=1) ²
SetAddAttributes/AddAttributes-10     0.000 ± ∞ ¹     0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
SimpleProcessorOnEmit-10              0.000 ± ∞ ¹     0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                                         ⁴                  +0.27%               ⁴
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal
³ need >= 4 samples to detect a difference at alpha level 0.05
⁴ summaries must be >0 to compute geomean

                                  │  ./old.txt  │              ./new.txt              │
                                  │  allocs/op  │  allocs/op   vs base                │
BatchProcessorOnEmit-10             0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
Processor/Simple-10                 1.000 ± ∞ ¹   1.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
Processor/Batch-10                  1.000 ± ∞ ¹   1.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
Processor/SetTimestampSimple-10     1.000 ± ∞ ¹   1.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
Processor/SetTimestampBatch-10      1.000 ± ∞ ¹   1.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
Processor/AddAttributesSimple-10    1.000 ± ∞ ¹   1.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
Processor/AddAttributesBatch-10     1.000 ± ∞ ¹   1.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
Processor/SetAttributesSimple-10    2.000 ± ∞ ¹   2.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
Processor/SetAttributesBatch-10     2.000 ± ∞ ¹   2.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
LoggerNewRecord/5_attributes-10     0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
LoggerNewRecord/10_attributes-10    4.000 ± ∞ ¹   4.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
LoggerEnabled-10                    0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
LoggerProviderLogger-10             1.000 ± ∞ ¹   1.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
WalkAttributes/1_attributes-10      0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
WalkAttributes/10_attributes-10     0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
WalkAttributes/100_attributes-10    0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
WalkAttributes/1000_attributes-10   0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
SetAddAttributes/SetAttributes-10   1.000 ± ∞ ¹   1.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
SetAddAttributes/AddAttributes-10   0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
SimpleProcessorOnEmit-10            0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                                       ³                +0.00%               ³
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal
³ summaries must be >0 to compute geomean
```

`sdk/metric/internal/aggregate`:

```
goos: darwin
goarch: arm64
pkg: go.opentelemetry.io/otel/sdk/metric/internal/aggregate
                                                                  │  ./old.txt   │              ./new.txt               │
                                                                  │    sec/op    │    sec/op     vs base                │
Prepend-10                                                          113.4µ ± ∞ ¹   113.4µ ± ∞ ¹       ~ (p=1.000 n=1) ²
Append-10                                                           26.28µ ± ∞ ¹   26.28µ ± ∞ ¹       ~ (p=1.000 n=1) ²
ExponentialHistogram/Int64/Cumulative/1/Measure-10                  90.61n ± ∞ ¹   88.46n ± ∞ ¹       ~ (p=1.000 n=1) ²
ExponentialHistogram/Int64/Cumulative/1/ComputeAggregation-10       156.2n ± ∞ ¹   136.9n ± ∞ ¹       ~ (p=1.000 n=1) ²
ExponentialHistogram/Int64/Cumulative/10/Measure-10                 1.088µ ± ∞ ¹   1.026µ ± ∞ ¹       ~ (p=1.000 n=1) ²
ExponentialHistogram/Int64/Cumulative/10/ComputeAggregation-10      431.6n ± ∞ ¹   419.1n ± ∞ ¹       ~ (p=1.000 n=1) ²
ExponentialHistogram/Int64/Cumulative/100/Measure-10                11.00µ ± ∞ ¹   10.40µ ± ∞ ¹       ~ (p=1.000 n=1) ²
ExponentialHistogram/Int64/Cumulative/100/ComputeAggregation-10     3.955µ ± ∞ ¹   3.954µ ± ∞ ¹       ~ (p=1.000 n=1) ²
ExponentialHistogram/Int64/Delta/1/Measure-10                       91.10n ± ∞ ¹   88.32n ± ∞ ¹       ~ (p=1.000 n=1) ²
ExponentialHistogram/Int64/Delta/1/ComputeAggregation-10            165.1n ± ∞ ¹   186.9n ± ∞ ¹       ~ (p=1.000 n=1) ²
ExponentialHistogram/Int64/Delta/10/Measure-10                      1.062µ ± ∞ ¹   1.038µ ± ∞ ¹       ~ (p=1.000 n=1) ²
ExponentialHistogram/Int64/Delta/10/ComputeAggregation-10           480.5n ± ∞ ¹   457.1n ± ∞ ¹       ~ (p=1.000 n=1) ²
ExponentialHistogram/Int64/Delta/100/Measure-10                     10.80µ ± ∞ ¹   10.36µ ± ∞ ¹       ~ (p=1.000 n=1) ²
ExponentialHistogram/Int64/Delta/100/ComputeAggregation-10          4.344µ ± ∞ ¹   4.301µ ± ∞ ¹       ~ (p=1.000 n=1) ²
ExponentialHistogram/Float64/Cumulative/1/Measure-10                89.94n ± ∞ ¹   87.71n ± ∞ ¹       ~ (p=1.000 n=1) ²
ExponentialHistogram/Float64/Cumulative/1/ComputeAggregation-10     124.4n ± ∞ ¹   124.2n ± ∞ ¹       ~ (p=1.000 n=1) ²
ExponentialHistogram/Float64/Cumulative/10/Measure-10               1.045µ ± ∞ ¹   1.029µ ± ∞ ¹       ~ (p=1.000 n=1) ²
ExponentialHistogram/Float64/Cumulative/10/ComputeAggregation-10    430.2n ± ∞ ¹   424.5n ± ∞ ¹       ~ (p=1.000 n=1) ²
ExponentialHistogram/Float64/Cumulative/100/Measure-10              10.63µ ± ∞ ¹   10.27µ ± ∞ ¹       ~ (p=1.000 n=1) ²
ExponentialHistogram/Float64/Cumulative/100/ComputeAggregation-10   4.026µ ± ∞ ¹   4.001µ ± ∞ ¹       ~ (p=1.000 n=1) ²
ExponentialHistogram/Float64/Delta/1/Measure-10                     91.85n ± ∞ ¹   87.60n ± ∞ ¹       ~ (p=1.000 n=1) ²
ExponentialHistogram/Float64/Delta/1/ComputeAggregation-10          154.8n ± ∞ ¹   151.3n ± ∞ ¹       ~ (p=1.000 n=1) ²
ExponentialHistogram/Float64/Delta/10/Measure-10                    1.147µ ± ∞ ¹   1.027µ ± ∞ ¹       ~ (p=1.000 n=1) ²
ExponentialHistogram/Float64/Delta/10/ComputeAggregation-10         461.1n ± ∞ ¹   462.4n ± ∞ ¹       ~ (p=1.000 n=1) ²
ExponentialHistogram/Float64/Delta/100/Measure-10                   10.46µ ± ∞ ¹   10.36µ ± ∞ ¹       ~ (p=1.000 n=1) ²
ExponentialHistogram/Float64/Delta/100/ComputeAggregation-10        4.508µ ± ∞ ¹   4.364µ ± ∞ ¹       ~ (p=1.000 n=1) ²
Histogram/Int64/Cumulative/1/Measure-10                             69.35n ± ∞ ¹   67.51n ± ∞ ¹       ~ (p=1.000 n=1) ²
Histogram/Int64/Cumulative/1/ComputeAggregation-10                  178.4n ± ∞ ¹   188.3n ± ∞ ¹       ~ (p=1.000 n=1) ²
Histogram/Int64/Cumulative/10/Measure-10                            840.3n ± ∞ ¹   816.6n ± ∞ ¹       ~ (p=1.000 n=1) ²
Histogram/Int64/Cumulative/10/ComputeAggregation-10                 630.2n ± ∞ ¹   613.9n ± ∞ ¹       ~ (p=1.000 n=1) ²
Histogram/Int64/Cumulative/100/Measure-10                           8.413µ ± ∞ ¹   8.279µ ± ∞ ¹       ~ (p=1.000 n=1) ²
Histogram/Int64/Cumulative/100/ComputeAggregation-10                6.180µ ± ∞ ¹   5.757µ ± ∞ ¹       ~ (p=1.000 n=1) ²
Histogram/Int64/Delta/1/Measure-10                                  68.25n ± ∞ ¹   68.22n ± ∞ ¹       ~ (p=1.000 n=1) ²
Histogram/Int64/Delta/1/ComputeAggregation-10                       171.3n ± ∞ ¹   170.3n ± ∞ ¹       ~ (p=1.000 n=1) ²
Histogram/Int64/Delta/10/Measure-10                                 829.7n ± ∞ ¹   810.4n ± ∞ ¹       ~ (p=1.000 n=1) ²
Histogram/Int64/Delta/10/ComputeAggregation-10                      455.4n ± ∞ ¹   456.7n ± ∞ ¹       ~ (p=1.000 n=1) ²
Histogram/Int64/Delta/100/Measure-10                                8.365µ ± ∞ ¹   8.266µ ± ∞ ¹       ~ (p=1.000 n=1) ²
Histogram/Int64/Delta/100/ComputeAggregation-10                     3.803µ ± ∞ ¹   3.835µ ± ∞ ¹       ~ (p=1.000 n=1) ²
Histogram/Float64/Cumulative/1/Measure-10                           69.71n ± ∞ ¹   68.19n ± ∞ ¹       ~ (p=1.000 n=1) ²
Histogram/Float64/Cumulative/1/ComputeAggregation-10                182.6n ± ∞ ¹   168.8n ± ∞ ¹       ~ (p=1.000 n=1) ²
Histogram/Float64/Cumulative/10/Measure-10                          859.8n ± ∞ ¹   819.0n ± ∞ ¹       ~ (p=1.000 n=1) ²
Histogram/Float64/Cumulative/10/ComputeAggregation-10               607.1n ± ∞ ¹   616.8n ± ∞ ¹       ~ (p=1.000 n=1) ²
Histogram/Float64/Cumulative/100/Measure-10                         8.400µ ± ∞ ¹   8.267µ ± ∞ ¹       ~ (p=1.000 n=1) ²
Histogram/Float64/Cumulative/100/ComputeAggregation-10              6.122µ ± ∞ ¹   6.247µ ± ∞ ¹       ~ (p=1.000 n=1) ²
Histogram/Float64/Delta/1/Measure-10                                70.10n ± ∞ ¹   70.38n ± ∞ ¹       ~ (p=1.000 n=1) ²
Histogram/Float64/Delta/1/ComputeAggregation-10                     174.9n ± ∞ ¹   168.0n ± ∞ ¹       ~ (p=1.000 n=1) ²
Histogram/Float64/Delta/10/Measure-10                               826.5n ± ∞ ¹   815.5n ± ∞ ¹       ~ (p=1.000 n=1) ²
Histogram/Float64/Delta/10/ComputeAggregation-10                    451.6n ± ∞ ¹   454.9n ± ∞ ¹       ~ (p=1.000 n=1) ²
Histogram/Float64/Delta/100/Measure-10                              8.535µ ± ∞ ¹   8.228µ ± ∞ ¹       ~ (p=1.000 n=1) ²
Histogram/Float64/Delta/100/ComputeAggregation-10                   3.878µ ± ∞ ¹   3.827µ ± ∞ ¹       ~ (p=1.000 n=1) ²
LastValue/Int64/1/Measure-10                                        121.0n ± ∞ ¹   120.3n ± ∞ ¹       ~ (p=1.000 n=1) ²
LastValue/Int64/1/ComputeAggregation-10                             149.4n ± ∞ ¹   149.9n ± ∞ ¹       ~ (p=1.000 n=1) ²
LastValue/Int64/10/Measure-10                                       1.486µ ± ∞ ¹   1.469µ ± ∞ ¹       ~ (p=1.000 n=1) ²
LastValue/Int64/10/ComputeAggregation-10                            412.8n ± ∞ ¹   412.3n ± ∞ ¹       ~ (p=1.000 n=1) ²
LastValue/Int64/100/Measure-10                                      15.32µ ± ∞ ¹   15.02µ ± ∞ ¹       ~ (p=1.000 n=1) ²
LastValue/Int64/100/ComputeAggregation-10                           3.286µ ± ∞ ¹   3.295µ ± ∞ ¹       ~ (p=1.000 n=1) ²
LastValue/Float64/1/Measure-10                                      120.9n ± ∞ ¹   121.6n ± ∞ ¹       ~ (p=1.000 n=1) ²
LastValue/Float64/1/ComputeAggregation-10                           150.7n ± ∞ ¹   150.3n ± ∞ ¹       ~ (p=1.000 n=1) ²
LastValue/Float64/10/Measure-10                                     1.479µ ± ∞ ¹   1.478µ ± ∞ ¹       ~ (p=1.000 n=1) ²
LastValue/Float64/10/ComputeAggregation-10                          405.3n ± ∞ ¹   465.9n ± ∞ ¹       ~ (p=1.000 n=1) ²
LastValue/Float64/100/Measure-10                                    15.11µ ± ∞ ¹   14.99µ ± ∞ ¹       ~ (p=1.000 n=1) ²
LastValue/Float64/100/ComputeAggregation-10                         3.239µ ± ∞ ¹   3.316µ ± ∞ ¹       ~ (p=1.000 n=1) ²
LimiterAttributes-10                                                198.2n ± ∞ ¹   206.8n ± ∞ ¹       ~ (p=1.000 n=1) ²
Sum/Int64/Cumulative/1/Measure-10                                   122.6n ± ∞ ¹   122.5n ± ∞ ¹       ~ (p=1.000 n=1) ²
Sum/Int64/Cumulative/1/ComputeAggregation-10                        130.7n ± ∞ ¹   127.5n ± ∞ ¹       ~ (p=1.000 n=1) ²
Sum/Int64/Cumulative/10/Measure-10                                  1.459µ ± ∞ ¹   1.484µ ± ∞ ¹       ~ (p=1.000 n=1) ²
Sum/Int64/Cumulative/10/ComputeAggregation-10                       424.5n ± ∞ ¹   367.7n ± ∞ ¹       ~ (p=1.000 n=1) ²
Sum/Int64/Cumulative/100/Measure-10                                 14.76µ ± ∞ ¹   15.06µ ± ∞ ¹       ~ (p=1.000 n=1) ²
Sum/Int64/Cumulative/100/ComputeAggregation-10                      2.899µ ± ∞ ¹   2.967µ ± ∞ ¹       ~ (p=1.000 n=1) ²
Sum/Int64/Delta/1/Measure-10                                        120.0n ± ∞ ¹   122.1n ± ∞ ¹       ~ (p=1.000 n=1) ²
Sum/Int64/Delta/1/ComputeAggregation-10                             154.8n ± ∞ ¹   157.0n ± ∞ ¹       ~ (p=1.000 n=1) ²
Sum/Int64/Delta/10/Measure-10                                       1.485µ ± ∞ ¹   1.464µ ± ∞ ¹       ~ (p=1.000 n=1) ²
Sum/Int64/Delta/10/ComputeAggregation-10                            401.2n ± ∞ ¹   409.2n ± ∞ ¹       ~ (p=1.000 n=1) ²
Sum/Int64/Delta/100/Measure-10                                      15.10µ ± ∞ ¹   15.12µ ± ∞ ¹       ~ (p=1.000 n=1) ²
Sum/Int64/Delta/100/ComputeAggregation-10                           3.170µ ± ∞ ¹   3.263µ ± ∞ ¹       ~ (p=1.000 n=1) ²
Sum/Float64/Cumulative/1/Measure-10                                 120.1n ± ∞ ¹   121.5n ± ∞ ¹       ~ (p=1.000 n=1) ²
Sum/Float64/Cumulative/1/ComputeAggregation-10                      118.8n ± ∞ ¹   121.8n ± ∞ ¹       ~ (p=1.000 n=1) ²
Sum/Float64/Cumulative/10/Measure-10                                1.475µ ± ∞ ¹   1.463µ ± ∞ ¹       ~ (p=1.000 n=1) ²
Sum/Float64/Cumulative/10/ComputeAggregation-10                     366.3n ± ∞ ¹   372.3n ± ∞ ¹       ~ (p=1.000 n=1) ²
Sum/Float64/Cumulative/100/Measure-10                               15.26µ ± ∞ ¹   15.02µ ± ∞ ¹       ~ (p=1.000 n=1) ²
Sum/Float64/Cumulative/100/ComputeAggregation-10                    2.898µ ± ∞ ¹   3.016µ ± ∞ ¹       ~ (p=1.000 n=1) ²
Sum/Float64/Delta/1/Measure-10                                      120.0n ± ∞ ¹   121.2n ± ∞ ¹       ~ (p=1.000 n=1) ²
Sum/Float64/Delta/1/ComputeAggregation-10                           151.8n ± ∞ ¹   153.8n ± ∞ ¹       ~ (p=1.000 n=1) ²
Sum/Float64/Delta/10/Measure-10                                     1.476µ ± ∞ ¹   1.469µ ± ∞ ¹       ~ (p=1.000 n=1) ²
Sum/Float64/Delta/10/ComputeAggregation-10                          398.2n ± ∞ ¹   408.8n ± ∞ ¹       ~ (p=1.000 n=1) ²
Sum/Float64/Delta/100/Measure-10                                    14.82µ ± ∞ ¹   14.95µ ± ∞ ¹       ~ (p=1.000 n=1) ²
Sum/Float64/Delta/100/ComputeAggregation-10                         3.312µ ± ∞ ¹   3.287µ ± ∞ ¹       ~ (p=1.000 n=1) ²
Sum/Precomputed/Int64/Cumulative/1/Measure-10                       121.1n ± ∞ ¹   122.6n ± ∞ ¹       ~ (p=1.000 n=1) ²
Sum/Precomputed/Int64/Cumulative/1/ComputeAggregation-10            157.7n ± ∞ ¹   155.4n ± ∞ ¹       ~ (p=1.000 n=1) ²
Sum/Precomputed/Int64/Cumulative/10/Measure-10                      1.483µ ± ∞ ¹   1.470µ ± ∞ ¹       ~ (p=1.000 n=1) ²
…