@pellared 2024-10-28
Do any of the existing benchmark show the performance improvement? If so then it would be good to add `benchstat` results. If not then we should add benchmarks that can be used to showcase the improvement and used for regression.