@ash2k 2024-10-28
@pellared 

> Do any of the existing benchmark show the performance improvement?

No, doesn't look like it. I've added benchmarks for the packages I edited. I don't know, maybe there are benchmarks in other packages that would show the difference (unlikely?).

> If not then we should add benchmarks that can be used to showcase the improvement and used for regression.

I'm sorry, but I don't have enough time to make significant changes in this PR. Lots of other things to do...

I've added a changelog entry.