@pellared 2024-11-06
> @pellared
> 
> > Do any of the existing benchmark show the performance improvement?
> 
> No, doesn't look like it. I've added output of benchmarks for the packages I edited to the description. I don't know, maybe there are benchmarks in other packages that would show the difference (unlikely?).
> 
> > If not then we should add benchmarks that can be used to showcase the improvement and used for regression.
> 
> I'm sorry, but I don't have enough time to make significant changes in this PR. Lots of other things to do...
> 
> I've added a changelog entry.

Actually, I do not think any regular benchmarks would show the improvement as basically it is a kind of memory leak (so it is not e.g. an unnecessary heap allocation). However, the existing benchmarks shows that clearing does not introduce any overhead.